Fix imports, rename module, add tests
