+++ prompts.py
@@ -1,5 +1,5 @@
 import yaml
 def load_prompt(key):
-    with open("config/prompts.yaml", "r") as f:
+    with open("prompts.yaml", "r") as f:
         prompts = yaml.safe_load(f)
     return prompts[key]
