+++ requirements.txt
@@ -5,3 +5,5 @@ langchain-groq
 openai
 numpy
 PyYAML
+pytest
+
