+++ thought_seeds.py
@@ -1,7 +1,8 @@
-from config.prompts import load_prompt
+from prompts import load_prompt
 from langchain_groq import ChatGroq
 
-llm = ChatGroq(model="mixtral-8x7b-32768")  # You can also use OpenAI, Claude etc.
+# You could swap ChatGroq with another LLM provider (e.g., OpenAI or Anthropic's Claude).
+llm = ChatGroq(model="mixtral-8x7b-32768")
 
 def generate_thought_seeds(state):
     user_input = state["user_input"]
